backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.1.0
django-filter==23.4
drf-spectacular==0.26.5
gunicorn==21.2.0
psycopg2-binary==2.9.9
groq==1.0.0
python-dotenv==1.0.0
python-pptx==0.6.23
dj-database-url==2.1.0
whitenoise==6.6.0
requests==2.31.0

